no acceptor (port is in use or requires root privileges)]
This link is misleading because it tells you to look for sinatra, which only returns the process that looks for sinatra. It is much better to look for the process ruby, because it will find the zombie server that is blocking port 9393 and preventing you from booting up the server.